# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 WeChatFerry RPC 实现的微信接口，使用 Go 语言编写，无第三方运行时依赖，易于对接任意编程语言。",
        "title": "Wrest Chat Api",
        "contact": {
            "name": "WeChatRest",
            "url": "https://github.com/opentdp/wrest-chat"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v0.10.0"
    },
    "basePath": "/",
    "paths": {
...
        "/bot/llmodel/create": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BOT::大语言模型"
                ],
                "summary": "添加模型",
                "parameters": [
                    {
                        "description": "添加模型参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/llmodel.CreateParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/bot/llmodel/delete": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BOT::大语言模型"
                ],
                "summary": "删除模型",
                "parameters": [
                    {
                        "description": "删除模型参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/llmodel.DeleteParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/bot/llmodel/detail": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BOT::大语言模型"
                ],
                "summary": "获取模型",
                "parameters": [
                    {
                        "description": "获取模型参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/llmodel.FetchParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tables.LLModel"
                        }
                    }
                }
            }
        },
        "/bot/llmodel/list": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BOT::大语言模型"
                ],
                "summary": "模型列表",
                "parameters": [
                    {
                        "description": "获取模型列表参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/llmodel.FetchAllParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/tables.LLModel"
                            }
                        }
                    }
                }
            }
        },
        "/bot/llmodel/update": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BOT::大语言模型"
                ],
                "summary": "修改模型",
                "parameters": [
                    {
                        "description": "修改模型参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/llmodel.UpdateParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/wcf/disable_receiver": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::消息推送"
                ],
                "summary": "关闭推送消息到URL",
                "parameters": [
                    {
                        "description": "推送消息到URL参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.ReceiverRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wcfrest.CommonPayload"
                        }
                    }
                }
            }
        },
...
        "/wcf/enable_receiver": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::消息推送"
                ],
                "summary": "开启推送消息到URL",
                "parameters": [
                    {
                        "description": "推送消息到URL参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.ReceiverRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wcfrest.CommonPayload"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "llmodel.CreateParam": {
            "type": "object",
            "required": [
                "family",
                "mid",
                "model",
                "provider",
                "secret"
            ],
            "properties": {
                "endpoint": {
                    "type": "string"
                },
                "family": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "mid": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "rd": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
        "llmodel.DeleteParam": {
            "type": "object",
            "properties": {
                "mid": {
                    "type": "string"
                },
                "rd": {
                    "type": "integer"
                }
            }
        },
        "llmodel.FetchAllParam": {
            "type": "object",
            "properties": {
                "family": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        },
        "llmodel.FetchParam": {
            "type": "object",
            "properties": {
                "mid": {
                    "type": "string"
                },
                "rd": {
                    "type": "integer"
                }
            }
        },
        "llmodel.UpdateParam": {
            "type": "object",
            "required": [
                "family",
                "mid",
                "model",
                "provider",
                "secret"
            ],
            "properties": {
                "endpoint": {
                    "type": "string"
                },
                "family": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "mid": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "rd": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
...
        "tables.LLModel": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间戳",
                    "type": "integer"
                },
                "endpoint": {
                    "description": "仅 google 和 openai 支持自定义",
                    "type": "string"
                },
                "family": {
                    "description": "模型家族，用于在指令中描述模型类型",
                    "type": "string"
                },
                "level": {
                    "description": "等级，用于限制用户访问",
                    "type": "integer"
                },
                "mid": {
                    "description": "模型 Id，用于生成模型切换指令",
                    "type": "string"
                },
                "model": {
                    "description": "模型，必须和服务商对应",
                    "type": "string"
                },
                "provider": {
                    "description": "服务商 [google, openai, xunfei, baidu, tencent]",
                    "type": "string"
                },
                "rd": {
                    "description": "主键",
                    "type": "integer"
                },
                "secret": {
                    "description": "密钥，必须和服务商对应，多个字段按指定顺序填写并用逗号隔开",
                    "type": "string"
                },
                "updated_at": {
                    "description": "最后更新时间戳",
                    "type": "integer"
                }
            }
        },
...
        "wcfrest.CommonPayload": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "错误信息"
                },
                "result": {
                    "description": "返回结果",
                    "type": "string"
                },
                "success": {
                    "description": "是否成功",
                    "type": "boolean"
                }
            }
        },
...
        "wcfrest.ReceiverRequest": {
            "type": "object",
            "properties": {
                "url": {
                    "description": "接收推送消息的 url",
                    "type": "string"
                }
            }
        },
...
    }
}